scripts/gsuite/requirements.txt
google-api-client
google-auth-httplib2
google-auth-oauthlib
google-api-python-client